+++ packages/sdk-communication-layer/src/services/SocketService/ChannelManager/setupChannelListeners.test.ts
@@ -6,8 +6,8 @@ import { setupChannelListeners } from './setupChannelListeners';
 
 describe('setupChannelListeners', () => {
   let instance: SocketService;
-  let socket: Partial<Socket>;
-  let keyExchange: Partial<KeyExchange>;
+  let socket: jest.Mocked<Partial<Socket>>;
+  let keyExchange: jest.Mocked<Partial<KeyExchange>>;
 
   const socketEvents = [
     'error',
@@ -15,7 +15,6 @@ describe('setupChannelListeners', () => {
     'reconnect',
     'reconnect_error',
     'reconnect_failed',
-    'disconnect',
   ];
   const channelEvents = [
     'clients_connected',
@@ -27,7 +26,7 @@ describe('setupChannelListeners', () => {
   const keyExchangeEvents = [EventType.KEY_INFO, EventType.KEYS_EXCHANGED]; // Adjust with the exact number of events in keyExchangeEventListenerMap
 
   beforeEach(() => {
-    socket = { on: jest.fn() };
+    socket = { on: jest.fn(), io: { on: jest.fn() } as any }; // Cast to any
     keyExchange = { on: jest.fn() };
 
     instance = {
@@ -48,10 +47,7 @@ describe('setupChannelListeners', () => {
     setupChannelListeners(instance, 'testChannelId');
 
     socketEvents.forEach((event) => {
-      expect(socket.on).toHaveBeenCalledWith(
-        `${event}-testChannelId`,
-        expect.any(Function),
-      );
+      expect(socket.io?.on).toHaveBeenCalledWith(event, expect.any(Function));
     });
   });