+++ packages/sdk-react-ui/src/MetaMaskUIProvider.tsx
@@ -85,11 +85,15 @@ const WagmiWrapper = ({
 // Wrap around to make sure the actual provider is only called on client to prevent nextjs issues.
 export const MetaMaskUIProvider = ({
   children,
+  connectors,
+  networks,
   sdkOptions,
   debug,
 }: {
   children: React.ReactNode;
   sdkOptions: MetaMaskSDKOptions;
+  connectors?: Connector[];
+  networks?: Chain[];
   debug?: boolean;
 }) => {
   const [clientSide, setClientSide] = useState(false);
@@ -102,7 +106,7 @@ export const MetaMaskUIProvider = ({
     <>
       {clientSide ? (
         <MetaMaskProvider debug={debug} sdkOptions={sdkOptions}>
-          <WagmiWrapper debug={debug}>
+          <WagmiWrapper debug={debug} connectors={connectors} networks={networks}>
             {children}
           </WagmiWrapper>
         </MetaMaskProvider>