+++ packages/sdk-react-ui/tailwind.config.js
@@ -1,6 +1,10 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: ['./src/**/*.{js,jsx,ts,tsx}'],
+  prefix: 'tw-',
+  corePlugins: {
+    preflight: false, // disable to prevent conflict with dapps css
+  },
   theme: {
     extend: {
       colors: {