+++ packages/sdk/rollup.config.js
@@ -11,13 +11,15 @@ import globals from 'rollup-plugin-node-globals';
 const packageJson = require('./package.json');
 
 const listDepForRollup = ['@react-native-async-storage/async-storage'];
+const webExternalDeps = [...listDepForRollup, 'qrcode-terminal'];
+const rnExternalDeps = [...listDepForRollup, 'qrcode-terminal'];
 
 /**
  * @type {import('rollup').RollupOptions}
  */
 const config = [
   {
-    external: listDepForRollup,
+    external: webExternalDeps,
     input: 'src/index.ts',
     output: [
       {
@@ -59,7 +61,7 @@ const config = [
     ],
   },
   {
-    external: listDepForRollup,
+    external: rnExternalDeps,
     input: 'src/index.ts',
     output: [
       {